Refactor offline mode and prompts
