+++ backend/tools/slides.py
@@ -12,233 +12,13 @@ from utils import extract_json_from_text
 from models import ResearchData, SlidePresentation, Slide
 from tools.slide_config import SLIDE_TYPES, PRESENTATION_STRUCTURE
 from tools.logo_fetcher import search_logo, download_logo
+from offline_responses.slides import generate_offline_slides
+from default_prompts import DEFAULT_SLIDES_PROMPT
+from prompts import get_prompt
 
 # Check for offline mode
 OFFLINE_MODE = os.environ.get("POWERIT_OFFLINE", "0").lower() in {"1", "true", "yes"}
 
-def generate_offline_slides(research: ResearchData, target_slides: int = 10, author: str = None) -> dict:
-    """
-    Generate realistic offline slides response based on the research content.
-    This creates a dynamic response that better matches what the real API would return.
-    """
-    # Use default author if none provided
-    if author is None:
-        author = PRESENTATION_STRUCTURE.get("default_author", "AI Presenter")
-    
-    # Extract key information from research content
-    research_content = research.content if research and research.content else "Artificial Intelligence in Modern Business"
-    
-    # Try to extract a title from the research content
-    lines = research_content.split('\n')
-    title_candidates = []
-    
-    # Look for title-like content in the first few lines
-    for line in lines[:10]:
-        line = line.strip()
-        if len(line) > 10 and len(line) < 100 and not line.startswith('-') and not line.startswith('•'):
-            title_candidates.append(line)
-    
-    # Use the first good candidate or generate a generic title
-    if title_candidates:
-        presentation_title = title_candidates[0].rstrip('.').rstrip(':')
-    else:
-        presentation_title = "Business Presentation"
-    
-    # Extract key topics for sections
-    content_sections = []
-    current_section = []
-    
-    for line in lines:
-        line = line.strip()
-        if line and len(line) > 5:
-            # Check if this looks like a section header
-            if (line.isupper() or 
-                line.endswith(':') or 
-                line.startswith('#') or 
-                (len(line) < 50 and not line.startswith('-') and not line.startswith('•'))):
-                if current_section:
-                    content_sections.append(current_section)
-                    current_section = []
-                current_section.append(line.rstrip(':').rstrip('#').strip())
-            else:
-                current_section.append(line)
-    
-    if current_section:
-        content_sections.append(current_section)
-    
-    # If no sections found, create generic ones
-    if not content_sections:
-        content_sections = [
-            ["Introduction", "Overview of the topic", "Key concepts and definitions"],
-            ["Main Content", "Core information and insights", "Important details and analysis"],
-            ["Applications", "Practical applications", "Real-world examples and use cases"],
-            ["Conclusion", "Summary of key points", "Next steps and recommendations"]
-        ]
-    
-    # Generate slides dynamically
-    slides = []
-    
-    # 1. Welcome slide
-    slides.append({
-        "type": "Welcome",
-        "fields": {
-            "title": presentation_title,
-            "subtitle": "Professional Business Presentation",
-            "author": author
-        }
-    })
-    
-    # 2. Process sections and create content
-    section_count = 0
-    slide_count = 1  # Already have welcome slide
-    
-    for section_data in content_sections[:4]:  # Limit to 4 sections to keep reasonable length
-        section_title = section_data[0] if section_data else f"Section {section_count + 1}"
-        section_content = section_data[1:] if len(section_data) > 1 else ["Content for this section"]
-        
-        # Add section divider
-        slides.append({
-            "type": "Section", 
-            "fields": {
-                "title": section_title
-            }
-        })
-        slide_count += 1
-        section_count += 1
-        
-        # Add 2-3 content slides per section with variety
-        slides_in_section = 0
-        max_slides_per_section = min(3, max(1, (target_slides - len(content_sections) - 1) // len(content_sections)))
-        
-        for i, content_item in enumerate(section_content[:max_slides_per_section]):
-            if slide_count >= target_slides:
-                break
-                
-            # Vary slide types for visual interest
-            if slides_in_section == 0 and len(content_item) > 20:
-                # First slide in section - use ContentImage for visual appeal
-                slides.append({
-                    "type": "ContentImage",
-                    "fields": {
-                        "title": f"{section_title} Overview",
-                        "subtitle": "Key Points and Insights",
-                        "content": [
-                            content_item,
-                            f"Important aspects of {section_title.lower()}",
-                            "Supporting details and context"
-                        ],
-                        "image": f"Professional illustration representing {section_title.lower()} concepts and applications in business environment"
-                    }
-                })
-            elif slides_in_section == 1 and section_count <= 2:
-                # Second slide - use ImageFull for emphasis
-                slides.append({
-                    "type": "ImageFull",
-                    "fields": {
-                        "title": f"{section_title} in Detail",
-                        "image": f"Detailed visual representation of {content_item[:50]}... showing modern business applications and innovative solutions",
-                        "explanation": f"This image illustrates the key concepts and practical applications related to {section_title.lower()}, demonstrating real-world implementation and benefits"
-                    }
-                })
-            elif slides_in_section == 2 and section_count == 2:
-                # Third slide in second section - use 3Images for comparison
-                slides.append({
-                    "type": "3Images",
-                    "fields": {
-                        "title": f"Three Aspects of {section_title}",
-                        "image1": f"Modern technology solutions for {section_title.lower()} showing digital transformation",
-                        "image2": f"Business processes and workflows related to {section_title.lower()} in corporate environment", 
-                        "image3": f"Future trends and innovations in {section_title.lower()} with growth projections",
-                        "image1subtitle": "Technology Solutions",
-                        "image2subtitle": "Business Processes", 
-                        "image3subtitle": "Future Innovations"
-                    }
-                })
-            elif "company" in content_item.lower() or "business" in content_item.lower() or any(word in content_item.lower() for word in ["google", "microsoft", "amazon", "apple", "meta"]):
-                # Use ContentWithLogos when content mentions companies
-                slides.append({
-                    "type": "ContentWithLogos",
-                    "fields": {
-                        "title": f"Leading Companies in {section_title}",
-                        "content": [
-                            content_item,
-                            "Industry leaders and innovators",
-                            "Key players driving innovation",
-                            "Market leaders setting industry standards"
-                        ],
-                        "logo1": "Google",
-                        "logo2": "Microsoft", 
-                        "logo3": "Amazon"
-                    }
-                })
-            else:
-                # Default to regular Content slide
-                slides.append({
-                    "type": "Content",
-                    "fields": {
-                        "title": content_item if len(content_item) < 60 else f"{section_title} Details",
-                        "content": [
-                            content_item,
-                            f"Key insights about {section_title.lower()}",
-                            "Supporting information and context",
-                            "Practical implications and applications"
-                        ]
-                    }
-                })
-            
-            slides_in_section += 1
-            slide_count += 1
-    
-    # Add a conclusion slide if we have room
-    if slide_count < target_slides:
-        slides.append({
-            "type": "Content",
-            "fields": {
-                "title": "Thank You & Questions",
-                "content": [
-                    "Thank you for your attention",
-                    "Questions and discussion",
-                    "Contact information available",
-                    "Additional resources provided"
-                ]
-            }
-        })
-    
-    # Process logos just like in online mode
-    print(f"OFFLINE MODE: Processing logos for ContentWithLogos slides")
-    for i, slide in enumerate(slides):
-        slide_type = slide.get('type')
-        if slide_type == 'ContentWithLogos':
-            print(f"OFFLINE MODE: Processing ContentWithLogos slide {i}")
-            
-            # Fetch logos from worldvectorlogo.com (same as online mode)
-            for logo_field in ['logo1', 'logo2', 'logo3']:
-                if logo_field in slide["fields"] and slide["fields"][logo_field]:
-                    logo_term = slide["fields"][logo_field]
-                    print(f"OFFLINE MODE: Fetching logo for '{logo_term}'")
-                    
-                    try:
-                        success, result = download_logo(logo_term)
-                        if success:
-                            # Update the slide with the path to the downloaded logo
-                            slide["fields"][logo_field] = result
-                            print(f"OFFLINE MODE: Successfully fetched logo for '{logo_term}': {result}")
-                        else:
-                            print(f"OFFLINE MODE WARNING: Failed to fetch logo for '{logo_term}': {result}")
-                            # If logo fetch fails, keep the text name
-                    except Exception as e:
-                        print(f"OFFLINE MODE ERROR: Exception while fetching logo '{logo_term}': {str(e)}")
-
-    # Add simple speaker notes for each slide
-    for slide in slides:
-        title = slide.get('fields', {}).get('title', 'this slide')
-        slide.setdefault('fields', {})['notes'] = f"Speaker notes for {title}."
-
-    return {
-        "title": presentation_title,
-        "author": author,
-        "slides": slides
-    }
 
 async def generate_slides(research: ResearchData, target_slides: int = 10, author: str = None) -> SlidePresentation:
     """
@@ -293,82 +73,8 @@ async def generate_slides(research: ResearchData, target_slides: int = 10, autho
     print(f"DEBUG: Slide types included in prompt: {', '.join(prompt_slide_types)}")
     
     # Create the prompt for slide generation with updated format for the new Slide model
-    prompt = f"""You are tasked with creating a professional presentation based on the following research:
-
-{research_content}
-
-Review the research content carefully and follow these guidelines:
-
-1. If the research content contains specific instructions about:
-   - Exact slide content
-   - Specific number of slides
-   - Particular speaker notes
-   - Specific visual elements
-   - Slide organization or structure
-   - Any other precise details about the presentation
-
-   Follow these instructions closely as specified. Do not deviate too much from explicit instructions, only if you want to introduce more visual elements.
-
-2. If the research does not contain specific instructions:
-   Create approximately {target_slides} slides for a professional presentation organized by sections.
-   
-IMPORTANT: You MUST use MULTIPLE DIFFERENT slide types. DO NOT default to using mostly 'Content' slides.
-PRIORITIZE using 'ContentImage', 'ImageFull', and '3Images' slides wherever possible to make the presentation visually engaging.
-AIM for no more than 30-40% of the main content slides (excluding Welcome/TOC/Section) to be of the plain 'Content' type.
-
-REQUIRED SLIDE TYPES (these MUST be included):\
-- 'Welcome' as the FIRST slide with title, subtitle, and author
-- At least 2-3 'Section' slides to divide your content into logical sections
-- A mix of 'Content' and 'ContentImage' slides (prefer 'ContentImage')
-- At least one '3Images' or 'ImageFull' slide if the content suits visual presentation
-
-Available slide types:
-{slide_types_info}
-
-SPECIAL INSTRUCTIONS FOR 'ContentWithLogos' SLIDES:
-- Use the 'ContentWithLogos' slide type when discussing specific companies, products, technologies, tools, or services
-- You can include 1 to 3 logos per slide
-- For each logo, simply include the company or product name (e.g., "Amazon AWS", "Microsoft", "Google Cloud")
-- Do not include any URLs or placeholders; the actual logos will be fetched automatically
-
-The presentation MUST follow this structure:
-- First slide MUST be a 'Welcome' slide that includes the presentation title, subtitle, and "{author}" as the author
-- Content MUST be organized into sections, each STARTING with a 'Section' slide
-- Use 'ContentImage' slides for important points that would benefit from visual illustration
-- Use regular 'Content' slides for general material when you have too much text, and no suitable image (do not have too many slides with only text)
-- Use 'ImageFull' for topics that need visual emphasis
-- Use '3Images' for comparing multiple items visually
-- Use 'ContentWithLogos' whenever discussing specific companies, products, or technologies that have recognizable logos
-
-The slides should be organized in clear sections (like "Market Research", "Our Services", "Reference Cases", etc.)
-
-REMEMBER: Use a VARIETY of slide types, prioritizing visual elements over plain text slides.
-
-For each slide, ONLY provide the required fields based on its type, as follows, and always include a "notes" field with 1-2 sentences of speaker notes:
-- Welcome slides: title, subtitle, author
-- Section slides: title
-- Content slides: title, content (MUST be a JSON array of strings, each string a bullet point)
-- ContentImage slides: title, subtitle, image, content (MUST be a JSON array of strings, each string a bullet point)
-- ImageFull slides: title, image, explanation
-- 3Images slides: title, image1, image2, image3, image1subtitle, image2subtitle, image3subtitle
-- ContentWithLogos slides: title, content (MUST be a JSON array of strings), logo1, logo2 (optional), logo3 (optional)
-
-When there are images, do not put placeholders, put the explanation what should be on the image.
-For logos, simply put the company or product name (e.g., "AWS", "Microsoft", "Google") - the actual logo will be fetched.
-
-DO NOT include any fields not specifically listed for each slide type. For example, Welcome slides should NOT have a content field.
-
-Note: Include the "notes" field for each slide. Do not include any "visual_suggestions" field.
-
-Format your response as a valid JSON object where:
-1. The root object has "title", "author", and "slides" properties
-2. Each slide has "type" and "fields" properties
-3. The "fields" property contains only the fields allowed for that slide type, AND the "content" field, when present, MUST be a JSON array of strings.
-
-Make sure your JSON is valid with proper escaping of quotes and special characters. 
-Each slide should only have the 'type' and 'fields' properties.
-The 'fields' property should only contain the fields specified for that slide type.
-"""
+    prompt_template = await get_prompt("slides_prompt", DEFAULT_SLIDES_PROMPT)
+    prompt = prompt_template.format(research_content=research_content, target_slides=target_slides, slide_types_info=slide_types_info, author=author)
     print(f"DEBUG: Prompt length: {len(prompt)} characters")
 
     try:
